refactor: move Danish zone
